% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bref_boxscore.R
\name{get_bref_boxscore}
\alias{get_bref_boxscore}
\title{Get bref boxscore table}
\usage{
get_bref_boxscore(gamelink = NULL)
}
\arguments{
\item{gamelink}{link to boxscore page}
}
\value{
data.frame of boxscore
}
\description{
Get bref boxscore table
}
